v0.1.2a
A good chunk of notes have been added to the code to help everyone else
with which direction this app will be taken. If anybody has issue with
the direction the code is being taken, they can more easily tell the
rest of us what's going on and why they're unsatisfied with the
direction of the code before we can get too carried away with proramming
certain aspects of the app. Among that, there are a few other minor
changes:
• The design for the Equipment Management tab has been given radio
buttons to determine whether a piece of equipment is a vehicle or not
• The Employee Management page has been given a big overhaul in its
design, opting to vertically separate functions between projects and the
employees that work (or could work) in those projects.
• A method called displayMessage has been added to the main form to
allow for popups in the event that a user does not fill things out
correctly
• The Inventory_Report code has been given a method called
displayMessageYesNo; so far this is being used to ask the user whether
or not they'd like to save a document of the inventory report before
printing it.